analysis/pyproject.toml
[project]
name = "analysis"
version = "0.1.0"
description = "AI Privacy Tracker analysis pipeline"
requires-python = ">=3.12,<3.13"
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.7",
    "python-dateutil>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/analysis"]

[tool.hatch.build.targets.wheel.force-include]
"src/analysis/rules/keywords.yml" = "analysis/rules/keywords.yml"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra -q"
